Ajustando cobertura de teste

+++ src/Controllers/bankAccountController.js
@@ -50,6 +50,10 @@ const getAll = async (req, res) => {
 
 const getBankAccountbyId = async (req, res) => {
     try {
+        const { id } = req.params;
+        if (!id || !mongoose.Types.ObjectId.isValid(id)) {
+            return res.status(500).send({ error: "ID inválido ou ausente" });
+        }
         const bankAccount = await BankAccount.findById(req.params.id); // Buscando conta pelo ID
         if (!bankAccount) {
             return res.status(404).send({ message: "Conta não encontrada" }); // Enviando mensagem de erro se a conta não for encontrada
@@ -64,6 +68,12 @@ const getBankAccountbyId = async (req, res) => {
 
 const deleteBankAccount = async (req, res) => {
     try {
+
+        const { id } = req.params;
+
+        if (!id || !mongoose.Types.ObjectId.isValid(id)) {
+            return res.status(500).send({ error: "ID inválido ou ausente" });
+        }
         const bankAccount = await BankAccount.findByIdAndDelete(req.params.id); // Deletando conta pelo ID
         if (!bankAccount) {
             return res.status(404).send({ message: "Conta não encontrada" }); // Enviando mensagem de erro se a conta não for encontrada
@@ -79,8 +89,8 @@ const deleteBankAccount = async (req, res) => {
 const updateBankAccount = async (req, res) => {
     try {
         const { id } = req.params;
-        if (!mongoose.Types.ObjectId.isValid(id)) {
-            return res.status(400).send({ message: "ID inválido" });
+        if (!id || !mongoose.Types.ObjectId.isValid(id)) {
+            return res.status(500).send({ error: "ID inválido ou ausente" });
         }
 
         // Log dos dados recebidos
